+++ frontend/src/pages/RequestorPage/RequestPanel/PathParamForm/index.ts
@@ -1,0 +1,1 @@
+export { PathParamForm } from "./PathParamForm";
